Change the taxonomic status of data in a taxonomic database

Only one of taxon_id or sci_name needs to be entered. Either will work if as long as it makes a partial or full match to one row in the data.

Usage

dct_change_status(
  tax_dat,
  taxon_id = NULL,
  sci_name = NULL,
  new_status,
  usage_id = NULL,
  usage_name = NULL,
  clear_usage_id = grepl("accepted", new_status, ignore.case = TRUE),
  strict = FALSE,
  quiet = FALSE,
  args_tbl = NULL
)

Arguments

tax_dat

Dataframe; taxonomic database in Darwin Core format.

taxon_id

Character vector of length 1; taxonID for the entry to be changed.

sci_name

Character vector of length 1; scientificName for the entry to be changed.

new_status

Character vector of length 1; updated taxonomicStatus to use for the entry.

usage_id

Character vector of length 1; taxonID for accepted name if new status is synonym.

usage_name

Character vector of length 1; scientificName for accepted name if new status is synonym.

clear_usage_id

Logical vector of length 1; should acceptedNameUsageID be set to NA?. Default: TRUE if new_status is "accepted" (case insensitive).

strict

Logical vector of length 1; should taxonomic checks be run on the updated taxonomic database?

quiet

Logical vector of length 1; should warnings be silenced?

args_tbl

A dataframe including columns corresponding to one or more of the above arguments, except for tax_dat. In this case, the input taxonomic database will be modified sequentially over each row of input in args_tbl.

Value

Dataframe; taxonomic database in Darwin Core format

Details

usage_id or usage_name should be provided if the new status is a synonym; either will work as it makes a partial or full match to one row in the data.

Can either modify a single row in the input taxonomic database if each argument is supplied as a vector of length 1, or can apply a set of changes to the taxonomic database if the input is supplied as a dataframe.

Examples

# Swap the accepted / synonym status of
# Cephalomanes crassum (Copel.) M. G. Price
# and Trichomanes crassum Copel.
dct_filmies |>
  dct_change_status(
    sci_name = "Cephalomanes crassum (Copel.) M. G. Price",
    new_status = "synonym",
    usage_name = "Trichomanes crassum Copel."
  ) |>
  dct_change_status(
    sci_name = "Trichomanes crassum Copel.",
    new_status = "accepted"
  ) |>
  dct_validate(
    check_tax_status = FALSE,
    check_mapping_strict = FALSE,
    check_sci_name = FALSE)
#> # A tibble: 2,729 × 6
#>    taxonID  acceptedNameUsageID taxonomicStatus          taxon…¹ scien…² modif…³
#>    <chr>    <chr>               <chr>                    <chr>   <chr>   <chr>  
#>  1 54115096 NA                  accepted name            species Cephal… NA     
#>  2 54133783 NA                  accepted                 species Tricho… 2022-1…
#>  3 54115097 54133783            synonym                  species Cephal… 2022-1…
#>  4 54133784 54115098            synonym                  species Tricho… NA     
#>  5 54115098 NA                  accepted name            species Cephal… NA     
#>  6 54133785 54115099            synonym                  species Tricho… NA     
#>  7 54115099 NA                  provisionally accepted … species Cephal… NA     
#>  8 54133786 54115100            synonym                  species Cephal… NA     
#>  9 54133787 54115100            synonym                  species Cephal… NA     
#> 10 54133788 54115100            synonym                  species Cephal… NA     
#> # … with 2,719 more rows, and abbreviated variable names ¹​taxonRank,
#> #   ²​scientificName, ³​modified

# Sometimes changing one name will affect others, if they map
# to the new synonym
dct_change_status(
  tax_dat = dct_filmies |> head(),
  sci_name = "Cephalomanes crassum (Copel.) M. G. Price",
  new_status = "synonym",
  usage_name = "Cephalomanes densinervium (Copel.) Copel."
)
#> # A tibble: 6 × 6
#>   taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName  modif…¹
#>   <chr>    <chr>               <chr>           <chr>     <chr>           <chr>  
#> 1 54115096 NA                  accepted name   species   Cephalomanes a… NA     
#> 2 54133783 54115098            synonym         species   Trichomanes cr… 2022-1…
#> 3 54115097 54115098            synonym         species   Cephalomanes c… 2022-1…
#> 4 54133784 54115098            synonym         species   Trichomanes de… NA     
#> 5 54115098 NA                  accepted name   species   Cephalomanes d… NA     
#> 6 54133785 54115099            synonym         species   Trichomanes in… NA     
#> # … with abbreviated variable name ¹​modified

# Apply a set of changes
library(tibble)
updates <- tibble(
  sci_name = c("Cephalomanes atrovirens Presl",
    "Cephalomanes crassum (Copel.) M. G. Price"),
  new_status = "synonym",
  usage_name = "Trichomanes crassum Copel."
)
dct_filmies |>
  dct_change_status(args_tbl = updates) |>
  dct_change_status(sci_name = "Trichomanes crassum Copel.",
    new_status = "accepted")
#> # A tibble: 2,729 × 6
#>    taxonID  acceptedNameUsageID taxonomicStatus          taxon…¹ scien…² modif…³
#>    <chr>    <chr>               <chr>                    <chr>   <chr>   <chr>  
#>  1 54115096 NA                  accepted name            species Cephal… NA     
#>  2 54133783 NA                  accepted                 species Tricho… 2022-1…
#>  3 54115097 54133783            synonym                  species Cephal… 2022-1…
#>  4 54133784 54115098            synonym                  species Tricho… NA     
#>  5 54115098 NA                  accepted name            species Cephal… NA     
#>  6 54133785 54115099            synonym                  species Tricho… NA     
#>  7 54115099 NA                  provisionally accepted … species Cephal… NA     
#>  8 54133786 54115100            synonym                  species Cephal… NA     
#>  9 54133787 54115100            synonym                  species Cephal… NA     
#> 10 54133788 54115100            synonym                  species Cephal… NA     
#> # … with 2,719 more rows, and abbreviated variable names ¹​taxonRank,
#> #   ²​scientificName, ³​modified