Modify row(s) of a taxonomic database

Modify one or more rows in a taxonomic database in Darwin Core (DWC) format.

Usage

dct_modify_row(
  tax_dat,
  taxon_id = NULL,
  sci_name = NULL,
  tax_status = NULL,
  usage_id = NULL,
  usage_name = NULL,
  clear_usage_id,
  clear_usage_name,
  fill_usage_name,
  remap_names,
  remap_variant,
  stamp_modified,
  strict,
  quiet,
  args_tbl = NULL,
  ...
)

Arguments

tax_dat

Dataframe; taxonomic database in DWC format.

taxon_id

Character or numeric vector of length 1; taxonID of the row to be modified (the selected row). Can also use taxonID.

sci_name

Character vector of length 1; scientificName of the row to be modified if taxon_id is NULL, OR the scientificName to assign to the selected row if taxon_id is provided (see Details). Can also use scientificName.

tax_status

Character vector of length 1; taxonomicStatus to assign to the selected row. Can also use taxonomicStatus.

usage_id

Character or numeric vector of length 1; acceptedNameUsageID to assign to the selected row. Can also use acceptedNameUsageID

usage_name

Character vector of length 1; acceptedNameUsage to assign to the selected row. Can also use acceptedNameUsage.

clear_usage_id

Logical vector of length 1; should acceptedNameUsageID of the selected row be set to NA if the word "accepted" is detected in tax_status (not case-sensitive)? Default TRUE.

clear_usage_name

Logical vector of length 1; should acceptedNameUsageID of the selected row be set to NA if the word "accepted" is detected in tax_status (not case-sensitive)? Default TRUE.

fill_usage_name

Logical vector of length 1; should the acceptedNameUsage of the selected row be set to the scientificName corresponding to its acceptedNameUsageID? Default TRUE.

remap_names

Logical vector of length 1; should the acceptedNameUsageID be updated (remapped) for rows with the same acceptedNameUsageID as the taxonID of the row to be modified? Default TRUE.

remap_variant

Same as remap_names, but applies specifically to rows with taxonomicStatus of "variant". Default FALSE.

stamp_modified

Logical vector of length 1; should the modified column of any newly created or modified row include a timestamp with the date and time of its creation/modification? Default TRUE.

strict

Logical vector of length 1; should taxonomic checks be run on the updated taxonomic database? Default FALSE..

quiet

Logical vector of length 1; should warnings be silenced? Default FALSE..

args_tbl

A dataframe including columns corresponding to one or more of the above arguments, except for tax_dat. In this case, the input taxonomic database will be modified sequentially over each row of input in args_tbl. Other DWC terms can also be included as additional columns, similar to using ... to modify a single row.

...

other DWC terms to modify, specified as sets of named values. Each element of the vector must have a name corresponding to a valid DWC term; see dct_terms.

Value

Dataframe; taxonomic database in DWC format

Details

Arguments taxon_id, sci_name, tax_status, usage_id, and usage_name are provided as convenience to avoid typing the longer "taxonomicID", "scientificName", "taxonomicStatus", "acceptedNameUsageID", and "acceptedNameUsage", respectively, but the latter may be used instead.

taxon_id is only used to identify the row(s) to modify and is not itself modified. sci_name can be used in the same way if taxon_id is not provided (as long as sci_name matches a single row). If both taxon_id and sci_name are provided, sci_name will be assigned to the scientificName of the row identified by taxon_id, replacing any value that already exists.

usage_id and usage_name must match existing values of acceptedNameUsageID and acceptedNameUsage in the input data (tax_dat). On default settings, either can be used and the other will be filled in automatically (fill_usage_id and fill_usage_name are both TRUE).

tax_status any other arguments provided that are DWC terms will be assigned to the selected row (i.e., they will modify the row).

If remap_names is TRUE (default) and usage_id (i.e., acceptedNameUsageID) is provided, any names that have an acceptedNameUsageID matching the taxonID of the selected row (i.e., synonyms of that row) will also have their acceptedNameUsageID replaced with the new acceptedNameUsageID. This behavior is not applied to names with taxonomicStatus of "variant" by default, but can be turned on for such names with remap_variant.

If clear_usage_id or clear_usage_name is TRUE and tax_status (or taxonomicStatus) includes the word "accepted", acceptedNameUsageID or acceptedNameUsage will be set to NA respectively, regardless of the values of usage_id, usage_name, or fill_usage_name.

Can either modify a single row in the input taxonomic database if each argument is supplied as a vector of length 1, or can apply a set of changes to the taxonomic database if the input is supplied as a dataframe via args_tbl.

Examples

# Swap the accepted / synonym status of
# Cephalomanes crassum (Copel.) M. G. Price
# and Trichomanes crassum Copel.
dct_filmies |>
  dct_modify_row(
    sci_name = "Cephalomanes crassum (Copel.) M. G. Price",
    tax_status = "synonym",
    usage_name = "Trichomanes crassum Copel."
  ) |>
  dct_modify_row(
    sci_name = "Trichomanes crassum Copel.",
    tax_status = "accepted"
  ) |>
  dct_validate(
    check_tax_status = FALSE,
    check_mapping_accepted_status = FALSE,
    check_sci_name = FALSE
  )
#> # A tibble: 2,451 × 6
#>    taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName modif…¹
#>    <chr>    <chr>               <chr>           <chr>     <chr>          <chr>  
#>  1 54115096 NA                  accepted        species   Cephalomanes … NA     
#>  2 54133783 NA                  accepted        species   Trichomanes c… 2022-1…
#>  3 54115097 54133783            synonym         species   Cephalomanes … 2022-1…
#>  4 54133784 54115098            synonym         species   Trichomanes d… NA     
#>  5 54115098 NA                  accepted        species   Cephalomanes … NA     
#>  6 54133786 54115100            synonym         species   Cephalomanes … NA     
#>  7 54133787 54115100            synonym         species   Cephalomanes … NA     
#>  8 54133788 54115100            synonym         species   Cephalomanes … NA     
#>  9 54133789 54115100            synonym         species   Cephalomanes … NA     
#> 10 54133790 54115100            synonym         species   Lacostea java… NA     
#> # … with 2,441 more rows, and abbreviated variable name ¹​modified
# Sometimes changing one name will affect others, if they map
# to the new synonym
dct_modify_row(
  tax_dat = dct_filmies |> head(),
  sci_name = "Cephalomanes crassum (Copel.) M. G. Price",
  tax_status = "synonym",
  usage_name = "Cephalomanes densinervium (Copel.) Copel."
)
#> # A tibble: 6 × 6
#>   taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName  modif…¹
#>   <chr>    <chr>               <chr>           <chr>     <chr>           <chr>  
#> 1 54115096 NA                  accepted        species   Cephalomanes a… NA     
#> 2 54133783 54115098            synonym         species   Trichomanes cr… 2022-1…
#> 3 54115097 54115098            synonym         species   Cephalomanes c… 2022-1…
#> 4 54133784 54115098            synonym         species   Trichomanes de… NA     
#> 5 54115098 NA                  accepted        species   Cephalomanes d… NA     
#> 6 54133786 54115100            synonym         species   Cephalomanes c… NA     
#> # … with abbreviated variable name ¹​modified
# Apply a set of changes
library(tibble)
updates <- tibble(
  sci_name = c(
    "Cephalomanes atrovirens Presl",
    "Cephalomanes crassum (Copel.) M. G. Price"
  ),
  tax_status = "synonym",
  usage_name = "Trichomanes crassum Copel."
)
dct_filmies |>
  dct_modify_row(args_tbl = updates) |>
  dct_modify_row(
    sci_name = "Trichomanes crassum Copel.",
    tax_status = "accepted"
  )
#> # A tibble: 2,451 × 6
#>    taxonID  acceptedNameUsageID taxonomicStatus taxonRank scientificName modif…¹
#>    <chr>    <chr>               <chr>           <chr>     <chr>          <chr>  
#>  1 54115096 54133783            synonym         species   Cephalomanes … NA     
#>  2 54133783 NA                  accepted        species   Trichomanes c… 2022-1…
#>  3 54115097 54133783            synonym         species   Cephalomanes … NA     
#>  4 54133784 54115098            synonym         species   Trichomanes d… NA     
#>  5 54115098 NA                  accepted        species   Cephalomanes … NA     
#>  6 54133786 54115100            synonym         species   Cephalomanes … NA     
#>  7 54133787 54115100            synonym         species   Cephalomanes … NA     
#>  8 54133788 54115100            synonym         species   Cephalomanes … NA     
#>  9 54133789 54115100            synonym         species   Cephalomanes … NA     
#> 10 54133790 54115100            synonym         species   Lacostea java… NA     
#> # … with 2,441 more rows, and abbreviated variable name ¹​modified