Check that a taxonomic database is correctly formatted

Stops with an error if any check fails. Most checks are geared towards being able to use the taxonomic database for taxonomic name resolution at the species level.

Usage

dct_validate(
  tax_dat,
  check_taxon_id = TRUE,
  check_mapping = TRUE,
  check_taxonomic_status = TRUE,
  check_acc_syn_diff = TRUE,
  check_col_names = TRUE
)

Arguments

tax_dat

Dataframe; taxonomic database in Darwin Core format

check_taxon_id

Logical; should all instances of taxonID be required to be non-missing and unique?

check_mapping

Logical; should all values of acceptedNameUsageID be required to map to the taxonID of an existing name?

check_taxonomic_status

Logical; should all taxonomic names be required to include the status of either "accepted" or "synonym"?

check_acc_syn_diff

Logical; should accepted names and synonyms be required to be different?

check_col_names

Logical; should all column names be required to be a valid Darwin Core term?

Value

Dataframe; taxonomic database in Darwin Core format. Will be the same as the input.

Details

For check_taxonomic_status and check_acc_syn_diff, "accepted" and "synonym" are determined by string matching of taxonomicStatus; so "provisionally accepted" is counted as "accepted", "ambiguous synonym" is counted as "synonym", etc (not case-sensitive).

For check_mapping, the acceptedNameUsageID must map to the taxonID of an existing name which itself does not have an acceptedNameUsageID (i.e., no multiple levels of mapping allowed).

Examples

dct_validate(dct_filmies)
#> # A tibble: 2,729 × 5
#>    taxonID  acceptedNameUsageID taxonomicStatus             taxonRank scientif…¹
#>    <chr>    <chr>               <chr>                       <chr>     <chr>     
#>  1 54115096 NA                  accepted name               species   Cephaloma…
#>  2 54133783 54115097            synonym                     species   Trichoman…
#>  3 54115097 NA                  accepted name               species   Cephaloma…
#>  4 54133784 54115098            synonym                     species   Trichoman…
#>  5 54115098 NA                  accepted name               species   Cephaloma…
#>  6 54133785 54115099            synonym                     species   Trichoman…
#>  7 54115099 NA                  provisionally accepted name species   Cephaloma…
#>  8 54133786 54115100            synonym                     species   Cephaloma…
#>  9 54133787 54115100            synonym                     species   Cephaloma…
#> 10 54133788 54115100            synonym                     species   Cephaloma…
#> # … with 2,719 more rows, and abbreviated variable name ¹​scientificName