Check that a taxonomic database is correctly formatted

Stops with an error if any check fails. Most checks are geared towards being able to use the taxonomic database for taxonomic name resolution at the species level.

Usage

dct_validate(
  tax_dat,
  check_taxon_id = TRUE,
  check_tax_status = TRUE,
  check_mapping = TRUE,
  check_mapping_strict = TRUE,
  check_sci_name = TRUE,
  check_status_diff = !check_sci_name,
  check_col_names = TRUE,
  valid_tax_status = Sys.getenv("VALID_TAX_STATUS", unset =
    "accepted, synonym, variant, NA"),
  on_success = "data",
  on_fail = "error"
)

Arguments

tax_dat

Dataframe; taxonomic database in Darwin Core format.

check_taxon_id

Logical; should all instances of taxonID be required to be non-missing and unique?

check_tax_status

Logical; should all taxonomic names be required to include a valid value for taxonomic status (by default, "accepted", "synonym", or "variant")?

check_mapping

Logical; should all values of acceptedNameUsageID be required to map to the taxonID of an existing name?

check_mapping_strict

Logical; should rules about mapping of variants and synonyms be enforced? (see Details)

check_sci_name

Logical; should all instances of scientificName be required to be non-missing and unique?

check_status_diff

Logical; should each scientific name be allowed to have only one taxonomic status? Default FALSE if check_sci_name is TRUE, since check_sci_name makes this check redundant.

check_col_names

Logical; should all column names be required to be a valid Darwin Core term?

valid_tax_status

Character vector of length 1; valid values for taxonomicStatus. Each value must be separated by a comma. Default "accepted, synonym, variant, NA". "NA" indicates that missing (NA) values are valid. Case-sensitive. Can also be set with the environmental variable "VALID_TAX_STATUS" (see Examples).

on_success

Character vector of length 1, either "logical" or "data". Describes what to do if the check passes. Default "data".

on_fail

Character vector of length 1, either "error" or "summary". Describes what to do if the check fails. Default "error".

Value

Depends on the result of the check and on values of on_fail and on_success:

If the check passes and on_success is "logical", return TRUE

If the check passes and on_success is "data", return the input dataframe

If the check fails and on_fail is "error", return an error

If the check fails and on_fail is "summary", issue a warning and return a dataframe with a summary of the reasons for failure

Details

For check_mapping_strict and check_status_diff, "accepted", "synonym", and "variant" are determined by string matching of taxonomicStatus; so "provisionally accepted" is counted as "accepted", "ambiguous synonym" is counted as "synonym", etc (case-sensitive).

For check_mapping_strict, the following rules are enforced:

Rows with taxonomicStatus of "synonym" (synonyms) must have an acceptedNameUsageID matching the taxonID of an accepted name ( taxonomicStatus of "accepted")

Rows with taxonomicStatus of "variant" (orthographic variants) must have an acceptedNameUsageID matching the taxonID of an accepted name or synonym (but not another variant)

Rows with taxonomicStatus of "accepted" must not have any value entered for acceptedNameUsageID

Rows with a value for acceptedNameUsageID must have a valid value for taxonomicStatus.

Examples

dct_validate(
  dct_filmies,
  check_sci_name = FALSE,
  valid_tax_status =
    paste("accepted name, ambiguous synonym, provisionally accepted name,",
          "synonym, variant")
)
#> # A tibble: 2,729 × 5
#>    taxonID  acceptedNameUsageID taxonomicStatus             taxonRank scientif…¹
#>    <chr>    <chr>               <chr>                       <chr>     <chr>     
#>  1 54115096 NA                  accepted name               species   Cephaloma…
#>  2 54133783 54115097            synonym                     species   Trichoman…
#>  3 54115097 NA                  accepted name               species   Cephaloma…
#>  4 54133784 54115098            synonym                     species   Trichoman…
#>  5 54115098 NA                  accepted name               species   Cephaloma…
#>  6 54133785 54115099            synonym                     species   Trichoman…
#>  7 54115099 NA                  provisionally accepted name species   Cephaloma…
#>  8 54133786 54115100            synonym                     species   Cephaloma…
#>  9 54133787 54115100            synonym                     species   Cephaloma…
#> 10 54133788 54115100            synonym                     species   Cephaloma…
#> # … with 2,719 more rows, and abbreviated variable name ¹​scientificName